migration: add new migration state wait-unplug
This patch is not ready for inclusion yet, I'm looking for
feedback/ideas on a particular problem. See below.

This patch adds a new migration state called wait-unplug.
It is entered after the SETUP state and will transition
into ACTIVE once all devices were succesfully unplugged from the guest.

So if a guest doesn't respond or takes long to honor the unplug request
the user will see the migration state 'wait-unplug'.

It adds a new function callback to VMStateDescription which is called for every
device that implements it and reports its device_pending status.

Now this loop in the migration thread:

while (qemu_savevm_state_guest_unplug_pending()) {
        continue;
}

clearly needs a condition to terminate after a while/or
a certain number of iterations. But I'm not sure what is a good
solution. How much waiting time is acceptable for a migration?

Signed-off-by: Jens Freimann <jfreimann@redhat.com>